+++ src/pages/guess-the-guild.tsx
@@ -2,8 +2,8 @@ import { Center, Text } from "@chakra-ui/react"
 import { ThemeProvider, useThemeContext } from "components/[guild]/ThemeContext"
 import Layout from "components/common/Layout"
 import { GameMenu } from "components/guess-the-guild/GameMenu"
-import { GuessGuildByLogo } from "components/guess-the-guild/GuessGuildByLogo"
-import { PairLogosToGuilds } from "components/guess-the-guild/PairLogosToGuilds"
+import { GuessGuildByLogo } from "components/guess-the-guild/gamemodes/GuessGuildByLogo"
+import { PairLogosToGuilds } from "components/guess-the-guild/gamemodes/PairLogosToGuilds"
 import { ReactNode, useState } from "react"
 
 function GuessTheGuild() {
